% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bitcoin_price.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Get the bitcoin price in US dollars}
\usage{
get_bitcoin_price()
}
\value{
number
}
\description{
Get the bitcoin price in US dollars
}
\examples{
get_bitcoin_price()
}
