% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_rates.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Look up the historical values of base currency in another currency.}
\usage{
get_exchange_rates(
  base = "USD",
  symbols = "HUF",
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  retried = 0
)
}
\arguments{
\item{base}{currency symbol}

\item{symbols}{currency symbol}

\item{start_date}{date}

\item{end_date}{date}

\item{retried}{number of times the function already failed}
}
\value{
\code{data.table} object
}
\description{
Look up the historical values of base currency in another currency.
}
\examples{
get_exchange_rates()
get_exchange_rates(base= 'USD', symbols='EUR')
get_exchange_rates(base= 'USD', symbols='GBP', start_date= "2021-04-24", end_date= "2021-05-10")
}
