% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchanges.R
\name{get_usdhuf}
\alias{get_usdhuf}
\title{Look up the values of a US dollar in Hungarian Forints}
\usage{
get_usdhuf(retried = 0)
}
\arguments{
\item{retried}{number of times the function already failed}
}
\value{
number
}
\description{
Look up the values of a US dollar in Hungarian Forints
}
