% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usdhufs.R
\name{get_usdhufs}
\alias{get_usdhufs}
\title{Look up the historical values of a US dollar in Hungarian Forints}
\usage{
get_usdhufs(start_date = Sys.Date() - 30, end_date = Sys.Date(), retried = 0)
}
\arguments{
\item{retried}{number of times the function already failed}
}
\value{
\code{data.table} object
}
\description{
Look up the historical values of a US dollar in Hungarian Forints
}
