% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usdhufs.R
\name{get_usdhufs}
\alias{get_usdhufs}
\title{Look up the historical values of a US dollar in Hungarian Forints}
\usage{
get_usdhufs(start_date = Sys.Date() - 30, end_date = Sys.Date(), retried = 0)
}
\arguments{
\item{start_date}{the first day of the time range you would like to query}

\item{end_date}{the last day of the time range you would like to query}

\item{retried}{number of times the function already failed}
}
\value{
\code{data.table} object
}
\description{
Look up the historical values of a US dollar in Hungarian Forints
}
\examples{
get_usdhufs()
get_usdhufs(start_date = "2021-04-11", end_date = "2021-05-11")
}
